CCMAIL: kde-imaging@kde.org
If you implement application for loading KIPI plugins, then please read
carefully, there are  a number of changes that you need to make to your
application.

For this check in I implemented the comments editor.
This resulted in a number of changes to other parts:

First, KimDaBa should not show this plugin, as it has better way built
in for setting comments - likewise other tools may not want to show it
because they might not have any way to support comments for images.

Therefore KIPI::PluginLoader::PluginLoader does now as first argument take
a QStringList of plugins not to load - the strings written there should be
the Name attribute of the desktop file for the given plugin.

Later today or tomorrow I anticipate to change the plugin loader class even more -
currently it has these instance variables with accessor methods for them:
   List         m_pluginList;
   QStringList  m_pluginNames;
   QStringList  m_libraryNames;

I need to add some extra information for each plugin, so I'd rather have
all the information in a class like:

class PluginInfo {
  QString name;
  QString comment;
  QString path;
  KIPI::Plugin* plugin;
  // etc
}
This will result in just one list of PluginInfo's that the host apps will
be able to get to, rather than a number of accessor methods as it is now.

Let me know if you have any objections.

Once I've fixed the above I'll make sure that the name is set during
loading, so that we do not have to specify the name twice - which is why
the KIPI::Plugin::id() was removed again (I added it a week ago)

Finally, the comments editor plugin needed to set the description, which is
why the ImageInfoShared classes that host apps implements has been changed
to include a setComment(), setName(), clearAttributes() and addAttributes()
methods.

Cheers
Jesper

svn path=/trunk/kdeextragear-libs-1/kipi-plugins/; revision=308013