Accordinly with Tom Albers, restoring temporally ' -nofork ' K3b invocation for next RC release. This is a temp issue. A better way will be to use the new (and complete) dcop interface available on k3b.
IMPORTANT : A new tool to ARCHIVE/RESTORE digikam database is planed by me and Tom for digiKam 0.8.1. It will be embeded on digiKam to support all database features. kipi CDArchiving tool will be obsolete for digikam in the future. We need a maintenor from kimdaba, gwenview or showimg project for this tool.

CCMAIL: kde-imaging@kde.org, digikam-devel@kde.org

svn path=/trunk/extragear/libs/kipi-plugins/; revision=457637