kipi-plugins from trunk : GPSSync tool : The GPS location editor dialognow use a dediced Google Maps to select the right place where have been taken the pictures.
Improvements :

- If the current picture already have GPS coordinates, the map will pan to the right place automaiticly at session startup!

- The GPS coordinates are automaticly captured by the dialog when the user right click on the map ! There is nothing to do manually to set the GPS location...

Angelo: with this plugin, i'm using a little php script to handle the google map with is interfaced with the plugin dialog ! This php script is hosted actually in a 3rd party digikam page. I would to host this page in the new kipi-plugins web project page. Can you help me ?

Noe : there is a copy of php script in svn.

CCMAIL: gerhard@kulzer.net, kde-imaging@ke.org

BUG: 133359
CCBUGS: 111560

svn path=/trunk/extragear/libs/kipi-plugins/; revision=589554