Fix login to Gallery 2.2.
This was a very annoying bug to track down but it essentially relates
to a bug in the generation of Cookies in Gallery 2.2.

Gallery actually generates 3 Set-Cookie headers in it's response which
seems excessive. The first one is a GALLERYSID=<blank> and the other
two set it to the correct value. The routine that parses cookies is
somewhat dumb and just passes any cookies it receives back to the
server on the next request. By passing the blank GALLERYSID, Gallery
ensured that it did not connect to your logged in session.

The cookie parsing code should be make more intellegent, but in the
mean time I've put in a very simple fix.

Testing and feedback appreciated.

This also applies the patch by Martin Kaufmann - thanks.
BUG: 142259

svn path=/trunk/extragear/libs/kipi-plugins/; revision=655231