added new options to set the behavior if unprocessed images are found.
Right now a popup window will appear all the time if unprocessed images
are found. This is annoying after a while, even for me :-)
The new settings will control the way the plugin handles those images:

Ask the user - same behavior as before
Keep         - do not remove unprocessed images
Remove       - well, this would remove unprocessed images :-)

This is really helpful when playing with the advanced settings and you
don't want to be bothered by the popup window all the time.

svn path=/trunk/extragear/graphics/kipi-plugins/; revision=890045