added a simple preview mode for the currently selected image
In preview mode, you are able to see the results of the correction for
the currently selected image. This is useful when you want to tweak the
settings in advanced mode.
There are three display modes:

Corrected image - the result of the correction
Original image  - the original image with the red-eyes
Correction mask - displays the mask used to correct the red-eye effect

You can switch between those modes by hovering over the image preview.
Clicking on the preview image will display the mask.
This information is also displayed as a WhatsThis message.

The preview is not yet complete, right now it will scale the image
without keeping aspect ratio, this will be fixed later.

svn path=/trunk/extragear/graphics/kipi-plugins/; revision=891016