GPSsync plugin uses common ImagesList widget now
Right now the plugin only uses the widget as it is, but I think this is
not good, it really doesn't help to maintain the plugin code.
It would be much better to merge gpslistviewitem and
gpstracklistviewitem into one common item (GPSListViewItem) and derive a
common ListView widget from ImagesList (GPSImageList), that can be used
in gpssyncdialog and gpstracklisteditdialog.

Since I'm not the author of the plugin, I might get things wrong here,
but for me those views and listitems are nearly the same.

For now I will leave it this way, if we want to merge these widgets, we
can do it later.

Andi

CCMAIL:digikam-devel@kde.org
CCMAIL:kde-imaging-@kde.org

svn path=/trunk/extragear/graphics/kipi-plugins/; revision=891658