Use struct as classpriv member
In C++ a struct is identical to a class except that the members of a struct have public visibility by default and the members of a class have private visibility by default.
To save me from typing "public:" all the time for class privs, I will use struct from now on in this plugin.

svn path=/trunk/extragear/graphics/kipi-plugins/; revision=909808