Extract the save methods into their own classes, this way it will be easier to add new correction modes.
The API will also be changed to have more then one detection /
correction method. Right now a HaarClassifier is used, which depends on
OpenCV. I am not too happy with the classifier so there is a chance that
a new detection method will be introduced someday. To make it
exchangeable, the plugin needs to be factorized a lot.
Since we will not be able to know anymore which libraries will be used
in the different detection methods, we need to unify the saving of the
results and let the appropriate detection methods take care of it.

This commit will prepare the plugin to have exchangeable save methods.
The SaveMethodFactory will generate a SaveMethod depending on the user
settings (read from the UI).
You provide the current image path to such a SaveMethod and it simply
returns the new path, depending on the methods used.
This path can than later be passed to the save method of the detection
system.

Right now the settings are static, so the subfolder and suffix name are
NOT read from the user settings.
For this to work, the next step will be to make RemovalSettings class a
singleton, to provide global access.

svn path=/trunk/extragear/graphics/kipi-plugins/; revision=912996