Quickfix: Do not create a list of comments/captions and angles when starting the slideshow, this will be very slow if you have more then 300 images selected. Instead read this info while images are loaded.
It is a quickfix because we still generate a fake comments and angle
list in slotSlideShow(), but with default values. I did it this way to
minimize code changes. In the future we might want to change the
slideshow classes instead.

Still the plugin wastes a lot of memory on startup, I have not found out
yet what produces this memory consumption.
Will need to do further checks.

CCBUG:182743

svn path=/trunk/extragear/graphics/kipi-plugins/; revision=1000030