Add possibilities to set a size for ImagesList widget and ImagesListView. The default iconsize was SizeLarge, but this may be to large for advancedslideshow. We need something to setup an default iconsize. ImagesListViewItem will get its iconsize from the associated view, and ImagesListView and ImagesList widget take an optional size parameter now.
This commit will also remove the creation of comments and tag
information, because the only plugin that needs it is advancedslideshow
and sendimages. If we add a lot of images to an ImagesList, it can take
a while to add this information.
If a plugin really needs this, it should call updateInformation() on the
ImagesListViewItem now, like sendimages does with this commit.

The next step will be to use ImagesList widget in advancedslideshow, to
factorize code.

CCBUG:182743

svn path=/trunk/extragear/graphics/kipi-plugins/; revision=1000476