apply patch #37922 by Martin EngelMann.
Comments from Martin :

"For this fix a hidden column is created in the BatchProcessImagesList that holds
the sort key for the BatchProcessImagesItems. This sort key is created based on 
the selected sorting criterium.

The already present part of the interface in BatchProcessImagesItem has been
replaced by get and set function for the sortkey. The column of the sortkey is
made available by the static function columnOfSortKey.

In batchprocessimageslist.cpp:44 the index of the column with the sort key is
used instead of calling the static function to avoid an include dependency with
batchprocessimagesitem.h. If this doesn't hurt. The function should be called  
instead of using the magic number.

I've deleted the already commented out function BatchProcessImagesItem::compare
because I don't see any further use of it."

BUGS: 212116


svn path=/trunk/extragear/graphics/kipi-plugins/; revision=1041769