Treat expiration time of 0 (infinity) properly
Bug introduced with the implementation for
http://www.digikam.org/drupal/node/610#comment-19928

(see also commit bf8d149a7148a787775b9052904eb51af34071e0)

I did not find a way to test without the offline_access (without
actually removing this flag from the code). I believe in the current
setup we'll always get infinite tokens. It'd be good to find a way
to make this optional, I've seen this in the past, but have currently
no idea how the requests would need to look like.