update gitversion.h properly
the file was never updated correctly, only when it has been manually removed
from the binary dir.
Also the version.h was never rebuild (I don't quite understand, why), even if
it was changed directly in the build directory. Somehow the dependency handling
was broken.

To fix this, the build directory needs to be completely removed after this
commit (make clean is not enough).