Avoid recompiling a lot of files when the git version changes
The problem is that we use the gitversion directly in the header file. Everytime the header changes, a lot of stuff needs to be recompiled.
Better yet, create the file "pluginsversion.cpp" and declare a function that will return the git version.
When the git version changes, only this file needs to be recompiled, the rest can be simply linked against it.

Again, the build directory must be recreated, otherwise it will not work.