Add kipi plugin for exporting to jAlbum
Create the minimal files for creating an album in jAlbum from the
selected images in digikam. The user must choose where to create
the directory for the album and the files are then created in
this directory. Finally jAlbum will be launched for the user
to make further alterations and actually create the album.
Requires a JVM and jAlbum both to be installed.

BUG: 316719
FIXED-IN: 3.2.0