Apply patch #82683 from Tim Newsome about to make smug export work with smugmug's new interface.
Changed most http_post() calls to get() calls. Apparently POST doesn't
work with the new interface albums.
Upgraded the interface we use from 1.2.0 to 1.2.2, since that's the
latest version in the 1.2 train.
Sort the list of albums that we receive from smugmug before presenting
it to the user.

BUGS: 323118
FIXES-IN: 4.0.0