Merge commit e02a738c15e1f46a6c1d3a4c569ff339df212401 from master to frameworks
googledrive: Escape photo title and description when inserting in JSON

Export to GDriver was failing when using with kipicmd because it sets
the description to 'Image located at "file:///path/to/image.jpg"'.
Because the double quotes were not escaped, the Google Drive server
received a broken JSON and complained about it.

CCBUG: 348278