Check for and use bundled JRE for jAlbum export if present
Add a check for the bundled Java Runtime Environment that is sometimes
shipped with jAlbum packages. This JRE will contain all required
dependencies for jAlbum when the system JRE may be missing some. If
found invoke jave from this bundled JRE.
If the bundled JRE cannot be found fallback to the system JRE which
was the behaviour before this commit.